+++ python/tvm/relax/__init__.py
@@ -16,8 +16,8 @@
 # under the License.
 # pylint: disable=invalid-name, wrong-import-position
 """The Relax IR namespace containing the IR, type, operator, builder, vm, etc."""
-from tvm.runtime import relax_vm as vm
-from tvm.runtime.relax_vm import VirtualMachine, VMInstrumentReturnKind
+from tvm.runtime import vm
+from tvm.runtime.vm import VirtualMachine, VMInstrumentReturnKind
 
 from .type_converter import args_converter
