+++ tests/python/relax/test_frontend_nn_extern_module.py
@@ -189,7 +189,7 @@ def test_extern_object():
         )
         _check_ir_equality(mod)
         mod = AttachExternModules(ext_mods)(mod)  # pylint: disable=not-callable
-        compiled = tvm.runtime.relax_vm.VirtualMachine(
+        compiled = tvm.runtime.vm.VirtualMachine(
             tvm.compile(mod, target="llvm"),
             device=tvm.cpu(),
         )
@@ -238,7 +238,7 @@ def test_extern_source():
     )
     _check_ir_equality(mod)
     mod = AttachExternModules(ext_mods)(mod)  # pylint: disable=not-callable
-    compiled = tvm.runtime.relax_vm.VirtualMachine(
+    compiled = tvm.runtime.vm.VirtualMachine(
         tvm.compile(mod, target="llvm"),
         device=tvm.cpu(),
     )
