+++ web/emcc/wasm_runtime.cc
@@ -60,16 +60,16 @@
 #include "ffi/src/ffi/traceback.cc"
 #include "src/runtime/memory/memory_manager.cc"
 #include "src/runtime/nvtx.cc"
-#include "src/runtime/relax_vm/attn_backend.cc"
-#include "src/runtime/relax_vm/builtin.cc"
-#include "src/runtime/relax_vm/bytecode.cc"
-#include "src/runtime/relax_vm/executable.cc"
-#include "src/runtime/relax_vm/kv_state.cc"
-#include "src/runtime/relax_vm/lm_support.cc"
-#include "src/runtime/relax_vm/ndarray_cache_support.cc"
-#include "src/runtime/relax_vm/paged_kv_cache.cc"
-#include "src/runtime/relax_vm/rnn_state.cc"
-#include "src/runtime/relax_vm/vm.cc"
+#include "src/runtime/vm/attn_backend.cc"
+#include "src/runtime/vm/builtin.cc"
+#include "src/runtime/vm/bytecode.cc"
+#include "src/runtime/vm/executable.cc"
+#include "src/runtime/vm/kv_state.cc"
+#include "src/runtime/vm/lm_support.cc"
+#include "src/runtime/vm/ndarray_cache_support.cc"
+#include "src/runtime/vm/paged_kv_cache.cc"
+#include "src/runtime/vm/rnn_state.cc"
+#include "src/runtime/vm/vm.cc"
 
 // --- Implementations of backend and wasm runtime API. ---
